# internal/handlers/handlers.go
package handlers

import (
	"Effective-Mobile-Test/internal/models"
	"Effective-Mobile-Test/internal/repository"
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"log/slog"
	"net/http"
	"strconv"
	"time"

	"github.com/google/uuid"
	"github.com/gorilla/mux"
)

type SubscriptionHandler struct {
	repo repository.RepositoryInterface
	log  *slog.Logger
}

func NewSubscriptionHandler(repo repository.RepositoryInterface, log *slog.Logger) *SubscriptionHandler {
	return &SubscriptionHandler{
		repo: repo,
		log:  log,
	}
}

func (h *SubscriptionHandler) RegisterRoutes(router *mux.Router) {
	router.HandleFunc("/subscriptions", h.CreateSubscriptionRecord).Methods("POST")
	router.HandleFunc("/subscriptions", h.ListSubsriptionRecords).Methods("GET")
	router.HandleFunc("/subscriptions/total-cost", h.CalculateSubscriptionCost).Methods("GET")

	router.HandleFunc("/subscriptions/{id:[0-9]+}", h.GetSubscriptionRecord).Methods("GET")
	router.HandleFunc("/subscriptions/{id:[0-9]+}", h.UpdateSubscriptionRecord).Methods("PUT")
	router.HandleFunc("/subscriptions/{id:[0-9]+}", h.PatchSubscriptionRecord).Methods("PATCH")
	router.HandleFunc("/subscriptions/{id:[0-9]+}", h.DeleteSubscriptionRecord).Methods("DELETE")
}

// @Summary Create new subscription record
// @Description Creates new subscription record
// @Tags subscriptions
// @Accept json
// @Produce json
// @Success 201 {object} models.SubscriptionResponse
// @Router /subscriptions [post]
func (h *SubscriptionHandler) CreateSubscriptionRecord(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
	defer cancel()

	if r.Header.Get("Content-Type") != "application/json" {
		http.Error(w, "Content-Type must be application/json", http.StatusUnsupportedMediaType)
		return
	}

	defer r.Body.Close()

	var createSubscription models.SubscriptionRequest

	body, err := io.ReadAll(r.Body)
	if err != nil {
		h.handleError(w, "Invalid request body", err, http.StatusBadRequest)
		return
	}

	err = json.Unmarshal(body, &createSubscription)
	if err != nil {
		h.handleError(w, "Invalid request body", err, http.StatusBadRequest)
		return
	}

	subscription, err := createSubscription.ToSubscription()
	if err != nil {
		h.handleError(w, "Invalid request body", err, http.StatusBadRequest)
		return
	}

	err = h.repo.Create(ctx, subscription)
	if err != nil {
		h.handleError(w, "Failed to create subscription record", err, http.StatusInternalServerError)
		return
	}

	h.log.Info("Subscription record created successfully", "ID", subscription.ID)

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)

	subscriptionResponse := subscription.ToResponse()

	data, err := json.Marshal(subscriptionResponse)
	if err != nil {
		h.log.Error("Failed to marshal subscription record", "error", err)
		return
	}
	w.Write(data)
}

// @Summary Get subscription record by ID
// @Description Gets subscription record by ID
// @Tags subscriptions
// @Produce json
// @Param id path int true "Subscription ID"
// @Success 200 {object} models.SubscriptionResponse
// @Router /subscriptions/{id} [get]
func (h *SubscriptionHandler) GetSubscriptionRecord(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
	defer cancel()

	defer r.Body.Close()

	vars := mux.Vars(r)

	id, err := strconv.Atoi(vars["id"])
	if err != nil {
		h.handleError(w, "Invalid id in request", err, http.StatusBadRequest)
		return
	}

	subscription, err := h.repo.GetByID(ctx, id)
	if err != nil {
		h.handleError(w, fmt.Sprintf("Failed to get subscription record with id: %d", id), err, http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	subscriptionResponse := subscription.ToResponse()
	data, err := json.Marshal(subscriptionResponse)
	if err != nil {
		h.handleError(w, "Failed to marshal response", err, http.StatusInternalServerError)
		return
	}
	w.Write(data)
	h.log.Info("Subscription record got successfully", "ID", subscriptionResponse.ID)
}

// @Summary List subscription records
// @Description Lists subscription records
// @Tags subscriptions
// @Produce json
// @Success 200 {array} models.SubscriptionResponse
// @Router /subscriptions [get]
func (h *SubscriptionHandler) ListSubsriptionRecords(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
	defer cancel()

	defer r.Body.Close()

	subscriptions, err := h.repo.List(ctx)
	if err != nil {
		h.handleError(w, "Failed to list subsription records", err, http.StatusInternalServerError)
		return
	}

	var response []*models.SubscriptionResponse
	for _, sub := range subscriptions {
		curr := sub.ToResponse()
		response = append(response, curr)
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
	h.log.Info("Subscription records listed successfully", "amount", len(response))
}

// @Summary Update subscription recored by ID
// @Description Makes full update of subscription recored by ID
// @Tags subscriptions
// @Accept json
// @Produce json
// @Param id path int true "Subscription ID"
// @Param subscription body models.SubscriptionRequest true "New data for subscription record"
// @Success 200 {object} models.SubscriptionResponse
// @Router /subscriptions/{id} [put]
func (h *SubscriptionHandler) UpdateSubscriptionRecord(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
	defer cancel()

	defer r.Body.Close()

	if r.Header.Get("Content-Type") != "application/json" {
		http.Error(w, "Content-Type must be application/json", http.StatusUnsupportedMediaType)
		return
	}

	vars := mux.Vars(r)
	id, err := strconv.Atoi(vars["id"])
	if err != nil {
		h.handleError(w, "Invalid id in request", err, http.StatusBadRequest)
		return
	}

	var updateSubscription models.SubscriptionRequest

	body, err := io.ReadAll(r.Body)
	if err != nil {
		h.handleError(w, "Invalid request", err, http.StatusBadRequest)
		return
	}

	err = json.Unmarshal(body, &updateSubscription)
	if err != nil {
		h.handleError(w, "Ivnalid request", err, http.StatusBadRequest)
		return
	}

	subscription, err := updateSubscription.ToSubscription()
	if err != nil {
		h.handleError(w, "Invalid request", err, http.StatusBadRequest)
		return
	}
	subscription.ID = id

	updatedSubscription, err := h.repo.Update(ctx, subscription)
	if err != nil {
		h.handleError(w, "Failed to update subscription record", err, http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	subscriptionResponse := updatedSubscription.ToResponse()
	data, err := json.Marshal(subscriptionResponse)
	if err != nil {
		h.handleError(w, "Failed to marshal updated subscription record", err, http.StatusInternalServerError)
		return
	}
	w.Write(data)

	h.log.Info("Subscription record updated successfully", "ID", id)
}

// @Summary Patch subscription record by ID
// @Description Makes partial update of subscription record by ID
// @Tags subscriptions
// @Accept json
// @Produce json
// @Param id path int true "Subscription ID"
// @Param subscription body models.SubscriptionRequest true "Data for partial updating subscription record"
// @Success 200 {object} models.SubscriptionResponse
// @Router /subscriptions/{id} [patch]
func (h *SubscriptionHandler) PatchSubscriptionRecord(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
	defer cancel()

	defer r.Body.Close()

	if r.Header.Get("Content-Type") != "application/json" {
		http.Error(w, "Content-Type must be application/json", http.StatusUnsupportedMediaType)
		return
	}

	vars := mux.Vars(r)
	id, err := strconv.Atoi(vars["id"])
	if err != nil {
		h.handleError(w, "Invalid request", err, http.StatusBadRequest)
		return
	}

	body, err := io.ReadAll(r.Body)
	if err != nil {
		h.handleError(w, "Invalid request", err, http.StatusBadRequest)
		return
	}

	var newSubscriptionRequest models.SubscriptionRequest
	err = json.Unmarshal(body, &newSubscriptionRequest)
	if err != nil {
		h.handleError(w, "Invalid request", err, http.StatusBadRequest)
		return
	}

	newSubscription, err := newSubscriptionRequest.ToSubscription()
	if err != nil {
		h.handleError(w, "Invalid request", err, http.StatusBadRequest)
		return
	}

	oldSubscription, err := h.repo.GetByID(ctx, id)
	if err != nil {
		h.handleError(w, "Failed to get subscription record by id", err, http.StatusInternalServerError)
		return
	}

	if newSubscription.ServiceName == "" {
		newSubscription.ServiceName = oldSubscription.ServiceName
	}

	if newSubscription.Price == 0 {
		newSubscription.Price = oldSubscription.Price
	}

	if newSubscription.UserID == uuid.Nil {
		newSubscription.UserID = oldSubscription.UserID
	}

	if newSubscription.StartDate.IsZero() {
		newSubscription.StartDate = oldSubscription.StartDate
	}

	if newSubscription.EndDate == nil {
		newSubscription.EndDate = oldSubscription.EndDate
	}

	newSubscription.ID = id

	updatedSubsription, err := h.repo.Update(ctx, newSubscription)
	if err != nil {
		fmt.Println(newSubscription)
		h.handleError(w, "Failed to update subscription record", err, http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	subscriptionResponse := updatedSubsription.ToResponse()
	data, err := json.Marshal(subscriptionResponse)
	if err != nil {
		h.handleError(w, "Failed to marshal response", err, http.StatusInternalServerError)
		return
	}

	w.Write(data)

	h.log.Info("Subscription record patched successfully", "id", id)
}

// @Summary Delete subscription record by ID
// @Description Deletes subscription record by ID
// @Tags subscriptions
// @Param id path int true "Subscription ID"
// @Success 204 "No content"
// @Router /subscriptions/{id} [delete]
func (h *SubscriptionHandler) DeleteSubscriptionRecord(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
	defer cancel()

	defer r.Body.Close()

	vars := mux.Vars(r)
	id, err := strconv.Atoi(vars["id"])
	if err != nil {
		h.handleError(w, "Invalid request to delete subscritption record", err, http.StatusBadRequest)
		return
	}

	err = h.repo.DeleteByID(ctx, id)
	if err != nil {
		h.handleError(w, "Failed to delete subscription record", err, http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusNoContent)
	h.log.Info("Subscription record deleted successfully", "id", id)
}

// @Summary Calculate subscriptin cost
// @Description Calculating subscription cost based on filtering parametres
// @Tags subscriptions
// @Produce json
// @Param start_date query string false "Start date of period (MM-YYYY)"
// @Param end_date query string false "End date of period (MM-YYYY)"
// @Param service_name query string false "Service name for filtering"
// @Param user_id query string fasle "User UUID for filtering"
// @Success 200 {object} models.SubscriptionCostResponse
// @Router /subscriptions/total-cost [get]
func (h *SubscriptionHandler) CalculateSubscriptionCost(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
	defer cancel()

	defer r.Body.Close()

	start_date := r.URL.Query().Get("start_date")
	end_date := r.URL.Query().Get("end_date")
	user_id := r.URL.Query().Get("user_id")
	service_name := r.URL.Query().Get("service_name")

	subscriptionCostRequest := models.SubscriptionCostRequest{
		StartDate:   start_date,
		EndDate:     end_date,
		UserID:      user_id,
		ServiceName: service_name,
	}

	subscriptionCost, err := subscriptionCostRequest.ToSubscriptionCost()
	if err != nil {
		h.handleError(w, "Invalid request", err, http.StatusBadRequest)
		return
	}

	if subscriptionCost.EndDate != nil && subscriptionCost.StartDate != nil && (*subscriptionCost).EndDate.Before(*subscriptionCost.StartDate) {
		h.handleError(w, "end date must be bigger than start date", errors.New("Invalid dates"), http.StatusBadRequest)
		return
	}

	cost, err := h.repo.CalculateSubscriptionCost(ctx, subscriptionCost)
	if err != nil {
		h.handleError(w, "Failed to calculate subscription cost", err, http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	subscriptionCostResponse := models.SubscriptionCostResponse{Cost: cost}
	data, err := json.Marshal(subscriptionCostResponse)
	if err != nil {
		h.handleError(w, "Failed to calculate subscription cost", err, http.StatusInternalServerError)
		return
	}
	w.Write(data)

	h.log.Info("Subscription cost calculated successfully", "cost", cost)
}

func (h *SubscriptionHandler) handleError(w http.ResponseWriter, message string, err error, status int) {
	http.Error(w, message, status)
	h.log.Error(message, "error", err)
}